purchase_monitoring/requirements.txt
Flask==2.2.2
flask_marshmallow==0.14.0
Flask_RESTful==0.3.9
Flask_SQLAlchemy==2.5.1
google_api_python_client==2.63.0
google_auth_oauthlib==0.5.3
pandas==1.3.3
protobuf==4.21.7
psycopg2==2.9.2
PyYAML==6.0
requests==2.26.0
SQLAlchemy==1.4.39
